+++ include/render/materials/ReflectiveMaterial.h
@@ -150,6 +150,9 @@ namespace render {
     MaterialBsdfSample sampleBsdf(const HitPoint& hitPoint, const Vector3d& wi,
                                   const Vector2d& sample) const override;
 
+    std::vector<MaterialBsdfSample> deltaBsdfSamples(const HitPoint& hitPoint,
+                                                     const Vector3d& wi) const override;
+
     double bsdfPdf(const HitPoint& hitPoint, const Vector3d& wi, const Vector3d& wo) const override;
 
     RasterRecursiveFallback rasterRecursiveFallback() const override {
@@ -162,6 +165,9 @@ namespace render {
     }
 
   protected:
+    MaterialBsdfSample reflectionDeltaBsdfSample(const HitPoint& hitPoint,
+                                                 const Vector3d& wi) const;
+
     render::PerfectSpecular m_reflectiveBRDF;
   };
 }